requirements.txt
requests
beautifulsoup4
PyYAML
python-dotenv
google-generativeai
httpx
youtube-transcript-api>=1.2.2
lxml
assemblyai
yt-dlp

# --- OPTIONAL: For X/Twitter Integration (implement fetcher in daily_run.py) ---
# tweepy

# --- OPTIONAL: For Instagram Integration (use with caution, may be unstable) ---
# instaloader
